+++ src/signdata/writer/ascii.go
