+++ src/signdata/writer/assembler.go
